add soft_nearest_neighbor_loss function
